refactor: 코드리뷰반영 - swagger data누락 & swagger 데코레이터명 수정
