+++ src/games/presentation/games.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { GamesService } from '../application/games.service';
 import { GetGameOptionsResponseDto } from './dto/get-game-options.dto';
-import { ApiGetGameOptions } from './swagger-decorators/get-game-options.decorator';
+import { ApiGetGameOptions } from './swagger-decorators/get-game-options-swagger.decorator';
 
 @ApiTags('Games')
 @Controller('games')
